pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imihigo-dhis2-tools"
version = "0.1.0"
description = "Setup DHIS2 Imihigo performance contract datasets for Eastern Province, Rwanda"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "questionary>=2.0",
    "requests>=2.31",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0", "responses>=0.25"]

[project.scripts]
imihigo = "imihigo_dhis2_tools.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/imihigo_dhis2_tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
